+++ functions/Get-WTKeybindings.ps1
@@ -13,7 +13,17 @@ Function Get-WTKeyBinding {
     $list = [System.Collections.Generic.List[Object]]::new()
 
     Write-Verbose "[$((Get-Date).TimeofDay)] Getting WindowsTerminal Appx package"
-    $install =(Get-AppxPackage -name Microsoft.WindowsTerminal).InstallLocation
+    <#
+    Need to get the correct application depending on whether running release or preview
+    #>
+    if ((Get-WTProcess | Where-Object {$_.name -eq 'WindowsTerminal'}).path -match 'preview') {
+        $pkg = "Microsoft.WindowsTerminalPreview"
+    }
+    else {
+        $pkg = "Microsoft.WindowsTerminal"
+    }
+    $install = (Get-AppxPackage -name $pkg).InstallLocation
+
     Write-Verbose "[$((Get-Date).TimeofDay)] Getting defaults.json file"
     $defaults = Join-Path -path $install -ChildPath defaults.json
 
@@ -23,7 +33,11 @@ Function Get-WTKeyBinding {
 
     #get the keybindings and add a property that indicates where the setting came from.
     Write-Verbose "[$((Get-Date).TimeofDay)] Parsing default keybindings"
-    $keys = $defaultsettings.keybindings |
+    <#
+    It looks like the json schema might be changing so I need to allow for name variations.
+    8/1/2020 jdh
+    #>
+    $keys = $defaultsettings | Select-Object -Expandproperty "*bindings" |
     parsesetting |
     Select-Object -Property *, @{Name = "Source"; Expression = {"Defaults"}}
 
@@ -31,29 +45,32 @@ Function Get-WTKeyBinding {
     #add the keybinding objects to the list
     $list.AddRange($keys)
 
-    Write-Verbose "[$((Get-Date).TimeofDay)] Getting user settings"
-    $settingsjson = "$ENV:Userprofile\AppData\Local\Packages\Microsoft.WindowsTerminal_8wekyb3d8bbwe\LocalState\settings.json"
-    if (Test-Path $settingsjson) {
+    $settingsjson = $global:wtsettingspath
+    Write-Verbose "[$((Get-Date).TimeofDay)] Getting user settings from $settingsjson"
+
+    if (Test-Path -path $settingsjson) {
+        Write-Verbose "[$((Get-Date).TimeofDay)] Converting content to json"
         $settings = Get-Content -path $settingsjson | Where-Object {$_ -notmatch "//"} | ConvertFrom-Json
-
+        #this might change and be bindings or keybindings
         #only process if there are keybindings
-        if ($settings.keybindings) {
-        $user = $settings.keybindings |
-        parsesetting |
-        Select-Object -Property *, @{Name = "Source"; Expression = {"Settings"}}
-        Write-Verbose "[$((Get-Date).TimeofDay)] Found $($keys.count) user keybindings"
+        $bind = $settings | Select-Object -ExpandProperty "*bindings"
+        if ($bind) {
+            $user =$bind |
+            parsesetting |
+            Select-Object -Property *, @{Name = "Source"; Expression = {"Settings"}}
+            Write-Verbose "[$((Get-Date).TimeofDay)] Found $($keys.count) user keybindings"
 
-        #if there is a duplicate key binding, remove the default
-        foreach ($item in $user) {
-            $existing = $list.where( {$_.keys -eq $item.keys})
-            if ($existing) {
+            #if there is a duplicate key binding, remove the default
+            foreach ($item in $user) {
+                $existing = $list.where( {$_.keys -eq $item.keys})
+                if ($existing) {
                     Write-Verbose "[$((Get-Date).TimeofDay)] Detected an override of $($existing| Out-String)"
-                [void]($list.Remove($existing))
+                    [void]($list.Remove($existing))
+                }
+                #add the entry
+                $list.Add($item)
             }
-            #add the entry
-            $list.Add($item)
-        }
-    } #if keybindings
+        } #if keybindings
     }
 
     Write-Verbose "[$((Get-Date).TimeofDay)] Formatting keybinding settings as $Format"